Update HUB75 panel dimensions to support larger sizes and adjust input limits in settings

+++ wled00/FX.h
@@ -971,8 +971,8 @@ class WS2812FX {
     struct Panel {
       uint16_t xOffset; // x offset relative to the top left of matrix in LEDs
       uint16_t yOffset; // y offset relative to the top left of matrix in LEDs
-      uint8_t  width;   // width of the panel
-      uint8_t  height;  // height of the panel
+      uint16_t width;   // width of the panel (uint16_t: allow flat HUB75 matrix > 255 wide, e.g. 5×64=320)
+      uint16_t height;  // height of the panel
       union {
         uint8_t options;
         struct {
